requirements.txt
aiogram
ffprobe-python